+++ src/main/java/com/klodnicki/Bike/v2/rest/controller/RentBikeController.java
@@ -7,7 +7,7 @@ import com.klodnicki.Bike.v2.DTO.rent.RentResponseDTO;
 import com.klodnicki.Bike.v2.model.RentRequest;
 import com.klodnicki.Bike.v2.model.entity.Bike;
 import com.klodnicki.Bike.v2.model.entity.ChargingStation;
-import com.klodnicki.Bike.v2.service.serviceInterface.GenericRentBikeService;
+import com.klodnicki.Bike.v2.service.interfacee.GenericRentBikeService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
