+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/DeleteService.java
@@ -1,4 +1,4 @@
-package com.klodnicki.Bike.v2.service.serviceInterface;
+package com.klodnicki.Bike.v2.service.interfacee;
 
 public interface DeleteService<ID> {
