+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/GenericChargingStationService.java
@@ -1,4 +1,4 @@
-package com.klodnicki.Bike.v2.service.serviceInterface;
+package com.klodnicki.Bike.v2.service.interfacee;
 
 import com.klodnicki.Bike.v2.DTO.station.StationForAdminResponseDTO;
 import com.klodnicki.Bike.v2.model.entity.ChargingStation;
