+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/RentService.java
@@ -1,4 +1,4 @@
-package com.klodnicki.Bike.v2.service.serviceInterface;
+package com.klodnicki.Bike.v2.service.interfacee;
 
 public interface RentService<T, G> {
     T rent(G obj);
