+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/ReturnService.java
@@ -1,4 +1,4 @@
-package com.klodnicki.Bike.v2.service.serviceInterface;
+package com.klodnicki.Bike.v2.service.interfacee;
 
 public interface ReturnService<G, ID, ID2, ID3> {
