+++ src/main/java/com/klodnicki/Bike/v2/service/interfacee/UpdateService.java
@@ -1,4 +1,4 @@
-package com.klodnicki.Bike.v2.service.serviceInterface;
+package com.klodnicki.Bike.v2.service.interfacee;
 
 public interface UpdateService<T, ID, G> {
